feat: add user validation on update/ delete video

+++ src/controllers/videoController.js
@@ -8,7 +8,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
 	const { id } = req.params;
-	const video = await Video.findById(id).populate('owner')
+	const video = await Video.findById(id).populate("owner");
 
 	if (!video) {
 		return res.status(404).render("404", { pageTitle: "Video not found" });
@@ -17,11 +17,22 @@ export const watch = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
-	const { id } = req.params;
+	const {
+		params: { id },
+		session: {
+			user: { _id: userId },
+		},
+	} = req;
 	const video = await Video.findById(id);
+
 	if (!video) {
 		return res.status(404).render("404", { pageTitle: "Video not found" });
 	}
+
+	if (userId != video.owner) {
+		return res.status(403).redirect("/");
+	}
+
 	return res.render("video/edit", {
 		pageTitle: `Edit ${video.title}`,
 		video,
@@ -30,28 +41,42 @@ export const getEdit = async (req, res) => {
 
 export const postEdit = async (req, res) => {
 	const { id } = req.params;
-	const { title, description, hashtags } = req.body;
-	try {
-		await Video.exists({ _id: id });
-	} catch {
+	const {
+		body: { title, description, hashtags },
+		session: {
+			user: { _id: userId },
+		},
+	} = req;
+
+	const video = await Video.findOne({ _id: id });
+
+	if (!video) {
 		return res.status(404).render("404", { pageTitle: "Video not found" });
 	}
+
+	if (userId != video.owner._id.toString()) {
+		return res.status(403).redirect("/");
+	}
+
 	await Video.findByIdAndUpdate(id, {
 		title,
 		description,
 		hashtags: Video.formatHashtags(hashtags),
 	});
+
 	return res.redirect(`../${id}`);
 };
+
 export const getUpload = (req, res) => {
 	return res.render("video/upload", { pageTitle: `Upload Video` });
 };
+
 export const postUpload = async (req, res) => {
 	const {
 		body: { title, description, hashtags },
 		file: { path },
 		session: {
-			user: { _id, videos },
+			user: { _id },
 		},
 	} = req;
 
@@ -75,7 +100,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
 	const { id } = req.params;
-	console.log({ id });
+	
+	const video = await Video.findOne({ _id: id });
+
+	if (!video) {
+		return res.status(404).render("404", { pageTitle: "Video not found" });
+	}
+
+	if (userId != video.owner._id.toString()) {
+		return res.status(403).redirect("/");
+	}
+	 
 	await Video.findByIdAndDelete(id);
 	return res.redirect("/");
 };
